requirements.txt
streamlit==1.26.0
opencv-python-headless==4.8.0.74
dlib-bin==19.24.0
scikit-learn==1.2.2
imutils==0.5.4
joblib==1.2.0
